+++ packages/components-js/tests/e2e/specs/segmented-control.e2e.ts
@@ -1,8 +1,11 @@
-import type { Page } from 'playwright';
 import { expect, test } from '@playwright/test';
+import { Components } from '@porsche-design-system/components';
+import type { Page } from 'playwright';
 import {
   addEventListener,
   getEventSummary,
+  getFormDataValue,
+  getHTMLAttributes,
   getLifecycleStatus,
   getOffsetWidth,
   getProperty,
@@ -17,28 +20,41 @@ import {
 const getHost = (page: Page) => page.locator('p-segmented-control');
 const getFirstItemHost = (page: Page) => page.locator('p-segmented-control-item').first();
 const getSecondItemHost = (page: Page) => page.locator('p-segmented-control-item:nth-child(2)');
-const getFirstItemButton = (page: Page) => page.locator('p-segmented-control-item button');
+const getFirstItemButton = (page: Page) => page.locator('p-segmented-control-item button').first();
 const getAllItemHosts = (page: Page) => page.locator('p-segmented-control-item').all();
+const getFieldset = (page: Page) => page.locator('fieldset');
 const getAllItemButtons = async (page: Page) =>
   Promise.all(
     (await getAllItemHosts(page)).map(async (x) =>
       (await x.evaluateHandle((x) => x.shadowRoot.querySelector('button'))).asElement()
     )
   );
+const getForm = (page: Page) => page.locator('form');
 
 const getFirstItemOffsetWidth = async (page: Page): Promise<number> => getOffsetWidth(getFirstItemHost(page));
 
-const initSegmentedControl = (page: Page, opts?: { amount?: number; value?: number }): Promise<void> => {
-  const { amount = 1, value } = opts || {};
+type InitOptions = {
+  props?: Components.PSegmentedControl;
+  isWithinForm?: boolean;
+  markupBefore?: string;
+  markupAfter?: string;
+  amount?: number;
+};
+
+const initSegmentedControl = (page: Page, opts?: InitOptions): Promise<void> => {
+  const { props = {}, isWithinForm = false, markupBefore = '', markupAfter = '', amount = 1 } = opts || {};
+
   const items = Array.from(Array(amount))
     .map((_, i) => `<p-segmented-control-item value="${i + 1}">Option ${i + 1}</p-segmented-control-item>`)
     .join('\n');
 
-  const content = `<p-segmented-control${value ? ` value="${value}"` : ''}>
+  const content = `<p-segmented-control ${getHTMLAttributes(props)}>
   ${items}
 </p-segmented-control>`;
 
-  return setContentWithDesignSystem(page, content);
+  const markup = `${markupBefore}${content}${markupAfter}`;
+
+  return setContentWithDesignSystem(page, isWithinForm ? `<form onsubmit="return false;">${markup}</form>` : markup);
 };
 
 test.describe('width calculation', () => {
@@ -165,14 +181,14 @@ test.describe('events', () => {
   });
 
   test('should not trigger event if item is selected', async ({ page }) => {
-    await initSegmentedControl(page, { value: 1 });
+    await initSegmentedControl(page, { props: { value: 1 } });
     const host = getHost(page);
     const firstItemHost = getFirstItemHost(page);
-    const button = await getFirstItemButton(page);
+    const button = getFirstItemButton(page);
 
     await addEventListener(host, 'segmentedControlChange');
 
-    expect(await getProperty(firstItemHost, 'selected')).toBe(true);
+    expect(await getProperty<boolean>(firstItemHost, 'selected')).toBe(true);
 
     await button.click();
     expect((await getEventSummary(host, 'segmentedControlChange')).counter).toBe(0);
@@ -202,7 +218,7 @@ test.describe('keyboard', () => {
       `<a href="#">Some Link</a>
 <p-segmented-control>
    <p-segmented-control-item value="1">Option 1</p-segmented-control-item>
-   <p-segmented-control-item value="2" disabled="true">Option 1</p-segmented-control-item>
+   <p-segmented-control-item value="2" disabled="true">Option 2</p-segmented-control-item>
 </p-segmented-control>
 <a href="#">Some Link</a>`
     );
@@ -252,5 +268,157 @@ test.describe('lifecycle', () => {
     );
 
     expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(2);
+  });
+});
+
+test.describe('form', () => {
+  test('should include name & value in FormData submit', async ({ page }) => {
+    const name = 'name';
+    const value = '2';
+    await initSegmentedControl(page, {
+      props: { name, value },
+      isWithinForm: true,
+      markupAfter: '<button type="submit">Submit</button>',
+    });
+    const form = getForm(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await page.locator('button[type="submit"]').click();
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(value);
+  });
+
+  test('should have correct form value when changing value dynamically', async ({ page }) => {
+    const name = 'name';
+    const value = '2';
+    const newValue = '1';
+    await initSegmentedControl(page, {
+      props: { name, value },
+      isWithinForm: true,
+      markupAfter: '<button type="submit">Submit</button>',
+    });
+    const form = getForm(page);
+    const host = getHost(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await setProperty(host, 'value', newValue);
+    await waitForStencilLifecycle(page);
+
+    await page.locator('button[type="submit"]').click();
+
+    await expect(host).toHaveJSProperty('value', newValue);
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(newValue);
+  });
+
+  test('should include name & value in FormData submit if outside of form', async ({ page }) => {
+    const name = 'name';
+    const value = 'Hallo';
+    const formId = 'myForm';
+    await initSegmentedControl(page, {
+      props: { name, value, form: formId },
+      markupBefore: `<form id="myForm" onsubmit="return false;"><button type="submit">Submit</button></form>`,
+    });
+    const form = getForm(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await page.locator('button[type="submit"]').click();
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(value);
+  });
+
+  test('should reset segmented-control value to its initial value on form reset', async ({ page }) => {
+    const name = 'name';
+    const value = '2';
+    const newValue = '1';
+    const host = getHost(page);
+    await initSegmentedControl(page, {
+      props: { name, value },
+      amount: 3,
+      isWithinForm: true,
+      markupAfter: `
+        <button type="submit">Submit</button>
+        <button type="reset">Reset</button>
+      `,
+    });
+    const form = getForm(page);
+    const button = getFirstItemButton(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await button.click();
+    await waitForStencilLifecycle(page);
+
+    await expect(host).toHaveJSProperty('value', newValue);
+
+    await page.locator('button[type="reset"]').click();
+
+    await expect(host).toHaveJSProperty('value', value);
+
+    await page.locator('button[type="submit"]').click(); // Check if ElementInternal value was reset as well
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(value);
+  });
+
+  test('should disable segmented-control if within disabled fieldset', async ({ page }) => {
+    const name = 'name';
+    const value = 'Hallo';
+    const host = getHost(page);
+    await initSegmentedControl(page, {
+      props: { name, value },
+      isWithinForm: true,
+      markupBefore: `<fieldset disabled>`,
+      markupAfter: `</fieldset>`,
+    });
+
+    await expect(host).toHaveJSProperty('disabled', true);
+  });
+
+  test('should sync disabled state with fieldset when updated programmatically', async ({ page }) => {
+    await initSegmentedControl(page, {
+      isWithinForm: true,
+      markupBefore: `<fieldset disabled>`,
+      markupAfter: `</fieldset>`,
+    });
+    const host = getHost(page);
+    const fieldset = getFieldset(page);
+    await expect(fieldset).toHaveJSProperty('disabled', true);
+    await expect(host).toHaveJSProperty('disabled', true);
+
+    await setProperty(fieldset, 'disabled', false);
+    await waitForStencilLifecycle(page);
+
+    await expect(fieldset).toHaveJSProperty('disabled', false);
+    await expect(host).toHaveJSProperty('disabled', false);
+  });
+
+  test('should allow item clicks when the control is programmatically re-enabled after being disabled', async ({
+    page,
+  }) => {
+    await initSegmentedControl(page, { amount: 2, props: { disabled: true } });
+    const host = getHost(page);
+    const [button1, button2] = await getAllItemButtons(page);
+    await addEventListener(host, 'segmentedControlChange');
+    await expect(host).toHaveJSProperty('disabled', true);
+
+    await setProperty(host, 'disabled', false);
+    await waitForStencilLifecycle(page);
+
+    await expect(host).toHaveJSProperty('disabled', false);
+    await button1.click();
+    expect((await getEventSummary(host, 'segmentedControlChange')).counter).toBe(1);
+
+    await button2.click();
+    expect((await getEventSummary(host, 'segmentedControlChange')).counter).toBe(2);
   });
 });