+++ packages/components/src/components.d.ts
@@ -1659,6 +1659,18 @@ export namespace Components {
          */
         "columns"?: BreakpointCustomizable<SegmentedControlColumns>;
         /**
+          * Disables the segmented-control.
+         */
+        "disabled"?: boolean;
+        /**
+          * The id of a form element the segmented-control should be associated with.
+         */
+        "form"?: string;
+        /**
+          * The name of the segmented-control.
+         */
+        "name"?: string;
+        /**
           * Adapts the segmented-control color depending on the theme.
          */
         "theme"?: Theme;
@@ -1702,7 +1714,7 @@ export namespace Components {
          */
         "description"?: string;
         /**
-          * Disables the select
+          * Disables the select.
          */
         "disabled"?: boolean;
         /**
@@ -4890,6 +4902,18 @@ declare namespace LocalJSX {
          */
         "columns"?: BreakpointCustomizable<SegmentedControlColumns>;
         /**
+          * Disables the segmented-control.
+         */
+        "disabled"?: boolean;
+        /**
+          * The id of a form element the segmented-control should be associated with.
+         */
+        "form"?: string;
+        /**
+          * The name of the segmented-control.
+         */
+        "name"?: string;
+        /**
           * @deprecated since v3.0.0, will be removed with next major release, use `update` event instead. Emitted when selected element changes.
          */
         "onSegmentedControlChange"?: (event: PSegmentedControlCustomEvent<SegmentedControlUpdateEventDetail>) => void;
@@ -4941,7 +4965,7 @@ declare namespace LocalJSX {
          */
         "description"?: string;
         /**
-          * Disables the select
+          * Disables the select.
          */
         "disabled"?: boolean;
         /**