+++ packages/components/CHANGELOG.md
@@ -34,6 +34,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
     ([3618](https://github.com/porsche-design-system/porsche-design-system/pull/3618))
 - `Pin-Code`: Add `form` prop to explicitly associate the component with a form, even when it's not directly nested
   within it. ([#3588](https://github.com/porsche-design-system/porsche-design-system/pull/3588))
+- `Segmented-Control`: Use ElementInternals API and add `form` prop to explicitly associate the component with a form,
+  even when it's not directly nested within it.
+  ([#3614](https://github.com/porsche-design-system/porsche-design-system/pull/3614))
 
 #### Changed
 
@@ -50,6 +53,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - Partials: removed bundled `react/jsx-runtime` due to React 18/19 incompatibilities. When using `jsx` in the `format`
   option, it is necessary to have `react/jsx-runtime` as a dependency in the project included.
   ([#3613](https://github.com/porsche-design-system/porsche-design-system/pull/3613))
+- `Select`, `Multi-Select`: Ensure that dynamically changing the `disabled` property via `optgroups` persists the
+  `disabled` state for individual options within the group.
+  ([#3614](https://github.com/porsche-design-system/porsche-design-system/pull/3614))
 
 ### [3.21.0] - 2024-11-12